requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0.0

# Data sources
yfinance>=0.2.0
ccxt>=4.0.0
requests>=2.31.0
aiohttp>=3.8.0

# Database
sqlite3  # Built-in with Python
psycopg2-binary>=2.9.0  # For PostgreSQL support

# ML and analysis
ta>=0.10.0  # Technical analysis
plotly>=5.15.0  # Visualization
jupyter>=1.0.0  # For analysis notebooks
tensorflow>=2.13.0  # For LSTM neural networks
keras>=2.13.0  # High-level neural network API

# Async and utilities
asyncio  # Built-in with Python
logging  # Built-in with Python
pytest>=7.4.0  # Testing
pytest-asyncio>=0.21.0  # Async testing

# Development
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Linting